[all] refactor MCP project
split into individual module files to improve future code maintenance of the mcp server
